Fixed Xcode document selection
For real this time: the workflow selects the correct Xcode source document by matching its file name with the (hidden) name of the current window. This approach still works only with the primary editor, but at least it handles tabs correctly.

Also refactored the workflow to gather both the document path and the selected line numbers upfront.